% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexibleglmm-package.R
\docType{package}
\name{FlexibleGLMM-package}
\alias{FlexibleGLMM}
\alias{FlexibleGLMM-package}
\title{flexibleglmm: Flexible GLMM Toolbox}
\description{
A Shiny app to fit GLMM models using afex, lme4::glmer, and
nlme::lme, with SAS/SPSS-like output, distribution fitting,
diagnostics, and post-hoc (EMMs) comparisons.
}
\author{
\strong{Maintainer}: Puneet Talwar \email{talwar.puneet@gmail.com}

Authors:
\itemize{
  \item Puneet Talwar \email{talwar.puneet@gmail.com}
}

}
\keyword{internal}
